refactor: club import 경로 변경
