+++ src/main/java/com/mobble/mobbleserver/domain/like/clubLike/service/ClubLikeService.java
@@ -1,11 +1,11 @@
 package com.mobble.mobbleserver.domain.like.clubLike.service;
 
-import com.mobble.mobbleserver.domain.club.club.entity.Club;
-import com.mobble.mobbleserver.domain.club.club.validator.ClubValidator;
-import com.mobble.mobbleserver.domain.like.clubLike.entity.ClubLike;
-import com.mobble.mobbleserver.domain.like.clubLike.repository.ClubLikeRepository;
+import com.mobble.mobbleserver.domain.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.validator.ClubValidator;
 import com.mobble.mobbleserver.domain.like.baseLike.entity.LikeType;
 import com.mobble.mobbleserver.domain.like.baseLike.service.AbstractLikeService;
+import com.mobble.mobbleserver.domain.like.clubLike.entity.ClubLike;
+import com.mobble.mobbleserver.domain.like.clubLike.repository.ClubLikeRepository;
 import com.mobble.mobbleserver.domain.member.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
