+++ src/test/java/com/mobble/mobbleserver/domain/article/entity/ArticleTest.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.domain.article.entity;
 
-import com.mobble.mobbleserver.domain.club.club.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.entity.Club;
 import com.mobble.mobbleserver.domain.clubCategory.entity.ClubCategory;
 import com.mobble.mobbleserver.domain.member.entity.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
